Added serialization and documentation
Commented the code in order to make it more obvious what was being done
with the various bits of code related to plugins.

Added save functionality, allowing configuration to be saved according
to whatever is required.

Fixed the save functionality on the plugins so that they are saved every
second or so.